feat: PTPage 컴포넌트 구현 및 슬라이드 기능 추가, 기존 PT 페이지 삭제
